Disable smart ten-partitioning, use Double

+++ sliderules-hs/src/SlideRules/Types.hs
@@ -2,5 +2,4 @@
 {-# LANGUAGE DataKinds #-}
 module SlideRules.Types where
 
-import Data.Decimal
-type InternalFloat = Decimal
+type InternalFloat = Double
